+++ validate/validate_test.go
@@ -306,22 +306,6 @@ func TestDatabaseVersion(t *testing.T) {
 			wantErr: false,
 		},
 		{
-			name: "supported engine and version does not return error",
-			args: args{
-				e: "mysql",
-				v: "8.0",
-			},
-			wantErr: false,
-		},
-		{
-			name: "supported engine and version does not return error",
-			args: args{
-				e: "mysql",
-				v: "8",
-			},
-			wantErr: false,
-		},
-		{
 			name: "supported engine and unsupported version returns error",
 			args: args{
 				e: "mysql",